+++ src/main/java/the_fireplace/grandexchange/commands/CommandOpSellThis.java
@@ -42,7 +42,7 @@ public class CommandOpSellThis extends CommandBase {
                 if(!isValidRequest)
                     throw new CommandException(TranslationUtil.getRawTranslationString(((EntityPlayer) sender).getUniqueID(), "commands.ge.common.not_holding_anything"));
                 ItemStack selling = ((EntityPlayerMP) sender).getHeldItemMainhand().isEmpty() ? ((EntityPlayerMP) sender).getHeldItemOffhand() : ((EntityPlayerMP) sender).getHeldItemMainhand();
-                long price = parseLong(args[0]);
+                double price = parseDouble(args[0]);
                 if (price < 0)
                     throw new CommandException(TranslationUtil.getRawTranslationString(((EntityPlayerMP) sender).getUniqueID(), "commands.ge.common.invalid_price"));
