+++ src/main/java/the_fireplace/grandexchange/compat/TaxDistributor.java
@@ -3,5 +3,5 @@ package the_fireplace.grandexchange.compat;
 import java.util.UUID;
 
 public interface TaxDistributor {
-    void distributeTax(UUID player, long amount);
+    void distributeTax(UUID player, double amount);
 }
